\name{automagicPaper}
\alias{automagicPaper}
\title{Get paper data through a query}
\usage{
  automagicPaper(query, verbose = T)
}
\arguments{
  \item{query}{Preferably a unique paper title}
}
\value{
  paper data as a data.frame
}
\description{
  This function returns the top result from Microsoft
  Academic Search for a given query.
}

