\name{cube2}
\alias{cube2}
\docType{data}
\title{A pantriagonal magic cube}
\description{
A pantriagonal magic cube of order 4 originally due to Hendricks
}
\usage{data(cube2)}
\details{
  Meaning of \dQuote{pantriagonal} currently unclear
}
\source{
Hendricks
}
\examples{
data(cube2)
is.magichypercube(cube2)
is.perfect(cube2)
}
\keyword{datasets}
