\name{force.integer}
\alias{force.integer}
\title{Integerize array elements}
\description{
Returns an elementwise \code{as.integer}-ed array.  All magic squares
should have integer elements.
}
\usage{
force.integer(x)
}
\arguments{
  \item{x}{Array to be converted}
}
\author{Robin K. S. Hankin}
\keyword{array}
