\name{getCountryName}
\alias{getCountryName}
\title{Get full country name from ISO alpha codes
This function looks up fully spelled-out country name from the \code{cpw_meta} table using ISO alpha code.}
\usage{
  getCountryName(iso.code)
}
\arguments{
  \item{iso}{Either a ISO2 or ISO3 alpha code}
}
\value{
  Country name as a string
}
\description{
  Get full country name from ISO alpha codes This function
  looks up fully spelled-out country name from the
  \code{cpw_meta} table using ISO alpha code.
}

