\name{getISOs}
\alias{getISOs}
\title{Get a list of all ISO 3166-1 alpha codes}
\usage{
  getISOs(iso.length)
}
\arguments{
  \item{iso.length}{Either values 2 or 3 to indicate which
  set of alpha codes to return.}
}
\value{
  A character vector of alpha codes
}
\description{
  Retrieve either ISO 3166-1 alpha-2 or alpha-3 codes from
  \code{cpw_meta.countries} table
}

