% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic.R
\name{ggplot.magic}
\alias{ggplot.magic}
\title{Convert a MAGIC object to a data.frame for ggplot}
\usage{
\method{ggplot}{magic}(data, ...)
}
\arguments{
\item{data}{A fitted MAGIC object}

\item{...}{Arguments for ggplot()}
}
\description{
Passes the smoothed data matrix to ggplot
}
\examples{
if (reticulate::py_module_available("phate") && require(ggplot2)) {

# data(tree.data)
# We use a smaller tree to make examples run faster
data(tree.data.small)
phate.tree <- phate(tree.data.small$data)
ggplot(phate.tree, aes(x=PHATE1, y=PHATE2, color=tree.data.small$branches)) +
  geom_point()

}
}
