\name{hendricks}
\alias{hendricks}
\docType{data}
\title{A perfect magic cube due to Hendricks}
\description{
A perfect \eqn{8\times 8\times 8}{8*8*8} magic cube due to Hendricks}
\usage{data(hendricks)}
\source{\url{http://members.shaw.ca/hdhcubes/cube_perfect.htm#Definitions}}
\examples{
data(hendricks)
is.perfect(hendricks)
}
\keyword{datasets}
\keyword{array}
