\name{iso2}
\alias{iso2}
\title{Get a valid country ISO2 alpha code}
\usage{
  iso2(country)
}
\arguments{
  \item{country}{Either a valid country name, ISO2 alpha
  code, or ISO3 alpha code}
}
\value{
  Two-character ISO2 alpha country code
}
\description{
  This function looks up ISO2 alpha country code from the
  \code{cpw_meta.countries} table.
}

