\name{iso3}
\alias{iso3}
\title{Get a valid country ISO3 alpha code}
\usage{
  iso3(country)
}
\arguments{
  \item{country}{Either a valid country name, ISO2 alpha
  code, or ISO3 alpha code}
}
\value{
  Three-character ISO3 alpha country code
}
\description{
  This function looks up ISO3 alpha country code from the
  \code{cpw_meta} table.
}

