\name{magicExplain.character}
\alias{magicExplain.character}
\title{Explains a character string}
\usage{
  magicExplain.character(variable.name = NULL,
    database = NULL)
}
\description{
  Provide instructions when \code{magicExplain} is given a
  character object (taken to mean a variable in
  \code{cpw_meta}).
}

