\name{magicExplain}
\alias{magicExplain}
\title{Explanations by \code{magic}}
\usage{
  magicExplain(obj, alpha = 0.05)
}
\description{
  This function explains a lot of things. In other words,
  it is a generic despatch method that routes objects to
  the right method based on class.
}
\examples{
\dontrun{
magicExplain()
magicExplain("s002")
}
}

