\name{magicOverview}
\alias{magicOverview}
\title{Visual map of literature}
\usage{
  magicOverview(user = NULL, prune = 1)
}
\arguments{
  \item{user}{Optional vector of user names to filter
  citations based on who contributed them into the
  database.}
}
\description{
  This function queries the \code{cpw_litReview} to
  visualize a citation network between the papers in the
  database.
}

