\name{magicSynopsis}
\alias{magicSynopsis}
\title{Show what others have said about a paper}
\usage{
  magicSynopsis(citekey)
}
\arguments{
  \item{citekey}{The citation key of the paper to be
  summarized.}
}
\value{
  A vector of sentences
}
\description{
  This function prints sentences that authors have used in
  reference to the focal paper. In many ways, these
  sentences summarize important points about it.
}

