\name{Ollerenshaw}
\alias{Ollerenshaw}
\docType{data}
\title{A most perfect square due to Ollerenshaw}
\description{
A 12-by-12 most perfect square due to Ollerenshaw
}
\usage{data(Ollerenshaw)}
\source{
  \dQuote{Most perfect pandiagonal magic squares}, K. Ollerenshaw and D. Bree,
  1998, Institute of Mathematics and its applications
}
\examples{
data(Ollerenshaw)
is.mostperfect(Ollerenshaw)
}
\keyword{datasets}
