\name{perfectcube6}
\alias{perfectcube6}
\docType{data}
\title{A perfect cube of order 6}
\description{
A perfect cube of order 6 originally due to Trump
}
\usage{data(perfectcube6)}
\examples{
data(perfectcube6)
is.perfect(perfectcube6)
is.magichypercube(perfectcube6[2:5,2:5,2:5])
}
\keyword{datasets}
