% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_magic.R
\name{run_magic}
\alias{run_magic}
\title{Do the magic!}
\usage{
run_magic(data, t_diffusion, lib_size_norm = TRUE, log_transform = FALSE,
  pseudo_count = 0.1, npca = 20, k = 30, ka = 10, epsilon = 1,
  rescale_percent = 0)
}
\arguments{
\item{data}{the data in some format}

\item{t_diffusion}{diffusion time; suggested to start with 6 and increase it
up to 12.}

\item{lib_size_norm}{Default: TRUE.}

\item{log_transform}{Default: FALSE.}

\item{npca}{number of PCA components that should be used; default: 20.}

\item{k}{Number of nearest neighbors to use when running MAGIC. Default: 30.}

\item{ka}{kNN-autotune parameter for running MAGIC; default: 10}

\item{epsilon}{a value for the standard deviation of the kernel. Default: 1.
\code{epsilon = 0} is the uniform kernel.}

\item{rescale_percent}{To which percentile should the data be re-scaled.
Note: Do not set this higher than 0 if you also want to log-transform.
Default: 0.}

\item{pseudo-count}{A number indicating how much should be added to avoid
log-transformation of zeros. Default: 0.1.}
}
