% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic.R
\name{summary.magic}
\alias{summary.magic}
\title{Summarize a MAGIC object}
\usage{
\method{summary}{magic}(object, ...)
}
\arguments{
\item{object}{A fitted MAGIC object}

\item{...}{Arguments for summary()}
}
\description{
Summarize a MAGIC object
}
\examples{
if (reticulate::py_module_available("phate")) {

# data(tree.data)
# We use a smaller tree to make examples run faster
data(tree.data.small)
phate.tree <- phate(tree.data.small$data)
summary(phate.tree)
## MAGIC embedding
## k = 5, alpha = NULL, t = 58
## Data: (3000, 100)
## Embedding: (3000, 2)

}
}
